#6871 : Home page Calendar widget are not showing Status value - Appu ======================================================================
Click on the plus(+) button on the Home page and create a widget for Calendar module with Status field selected. Now check the Status field in the widget ,it will not show any value although the it contains value if we will go to the detail view of the records.