#7172 : VTIGER_DB_VERSION test in index php wrong use of translation - Asha ============================================================================
Since user's current language can be detected only after checking for User's authentication, we will not be able to pick the right language for the translation of the message.

We will need to revamp the index.php file to handle the translation in right way.

For time being, we will remove the translation support for this and provide the message in English only.