Some Bug Fixes - Asha =============================
Fixed the issue with the background color and the Filters label color.
Fixed the issue with the image included in the footer.
Cron app_key removed from config.cron.php
Re-packaged the module packages changed recently
Reduced the size of subject input box of Send Email workflow task
Mail Scanner and Workflow cron job notification fixes
Moved Mail Scanner cron task as the last task on the  list
Fixed the position issue with 'Advanced Search' in IE
Fixed the position issue with Navigation bar of ListViews of all modules as well as Documents - On IE
Fixed the default frequencies for the cron jobs
Fixed the issue with Workflow includes, which is forcing Workflow cron job to fail
Fixed the issue with Custom Filter, throwing js error while creating Filter in Calendar module
Fixed the default frequencies for the cron jobs
Fixed the issue with Documents translation